+++ src/lab.rs
@@ -5,7 +5,6 @@
 
 use std::cmp::{max, min};
 use std::panic::resume_unwind;
-use std::path::Path;
 use std::sync::Mutex;
 use std::thread;
 use std::time::Instant;
@@ -213,29 +212,20 @@ fn test_scenario(
         .lock()
         .expect("lock output_dir to start scenario")
         .start_scenario(scenario)?;
-    scenario_output.message(&scenario.to_string())?;
-    console.scenario_started(
-        build_dir.path().as_ref(),
-        scenario,
-        scenario_output.open_log_read()?,
-    )?;
+    let dir = build_dir.path();
+    console.scenario_started(dir, scenario, scenario_output.open_log_read()?)?;
 
     let phases: &[Phase] = if options.check_only {
         &[Phase::Check]
     } else {
         &[Phase::Build, Phase::Test]
     };
-    let applied = scenario
-        .mutant()
-        .map(|mutant| {
-            // TODO: This is slightly inefficient as it computes the mutated source twice,
-            // once for the diff and once to write it out.
-            scenario_output.message(&format!("mutation diff:\n{}", mutant.diff()))?;
-            mutant.apply(build_dir)
-        })
-        .transpose()?;
-    let dir: &Path = build_dir.path().as_ref();
-    console.scenario_started(dir, scenario, scenario_output.open_log_read()?)?;
+    if let Some(mutant) = scenario.mutant() {
+        let mutated_code = mutant.mutated_code();
+        let diff = scenario.mutant().unwrap().diff(&mutated_code);
+        scenario_output.write_diff(&diff)?;
+        mutant.apply(build_dir, &mutated_code)?;
+    }
 
     let mut outcome = ScenarioOutcome::new(&scenario_output, scenario.clone());
     for &phase in phases {
@@ -244,7 +234,7 @@ fn test_scenario(
             Phase::Test => timeouts.test,
             Phase::Build | Phase::Check => timeouts.build,
         };
-        let phase_result = run_cargo(
+        match run_cargo(
             build_dir,
             jobserver,
             Some(test_packages),
@@ -253,15 +243,27 @@ fn test_scenario(
             &mut scenario_output,
             options,
             console,
-        )?;
-        let success = phase_result.is_success(); // so we can move it away
-        outcome.add_phase_result(phase_result);
-        console.scenario_phase_finished(dir, phase);
-        if !success {
-            break;
+        ) {
+            Ok(phase_result) => {
+                let success = phase_result.is_success(); // so we can move it away
+                outcome.add_phase_result(phase_result);
+                console.scenario_phase_finished(dir, phase);
+                if !success {
+                    break;
+                }
+            }
+            Err(err) => {
+                // Some unexpected internal error that stops the program.
+                if let Some(mutant) = scenario.mutant() {
+                    mutant.revert(build_dir)?;
+                    return Err(err);
+                }
+            }
         }
     }
-    drop(applied);
+    if let Some(mutant) = scenario.mutant() {
+        mutant.revert(build_dir)?;
+    }
     output_mutex
         .lock()
         .expect("lock output dir to add outcome")